+++ test/snapshot-order/backwards-compatibility.js
@@ -2,107 +2,7 @@ const test = require('@ava/test');
 const exec = require('../helpers/exec');
 const fs = require('fs');
 const path = require('path');
-
-test('snapshot files are independent of test resolution order', async t => {
-	const options = {
-		cwd: exec.cwd('intertest-order'),
-		env: {
-			AVA_FORCE_CI: 'not-ci'
-		}
-	};
-
-	const snapshotPath = path.join(options.cwd, 'test.js.snap');
-
-	// Schedule snapshot cleanup
-	t.teardown(() => {
-		fs.unlinkSync(snapshotPath);
-		fs.unlinkSync(path.join(options.cwd, 'test.js.md'));
-	});
-
-	// Run, updating snapshots.
-	await exec.fixture(['test.js', '--update-snapshots'], options);
-
-	// Read the resulting file
-	const snapshot = fs.readFileSync(snapshotPath);
-
-	// Run in reversed order, updating snapshots.
-	await exec.fixture(['test.js', '--update-snapshots'], {
-		...options,
-		env: {
-			INTERTEST_ORDER_REVERSE: 'true',
-			...options.env
-		}
-	});
-
-	// Read the resulting file
-	const snapshotReversed = fs.readFileSync(snapshotPath);
-
-	// Compare snapshots
-	t.deepEqual(snapshot, snapshotReversed);
-});
-
-function getSnapshotIds(report) {
-	function * matchAll(string, regexp) {
-		let match;
-		while ((match = regexp.exec(string)) !== null) {
-			yield match;
-		}
-	}
-
-	const ids = [...matchAll(report, /'index: ([-.\d]+)'/g)].map(match => Number(match[1]));
-
-	return ids;
-}
-
-test('snapshot reports are sorted in declaration order', async t => {
-	const options = {
-		cwd: exec.cwd('report-declaration-order'),
-		env: {
-			AVA_FORCE_CI: 'not-ci'
-		}
-	};
-
-	// Scehdule snapshot cleanup
-	t.teardown(() => {
-		fs.unlinkSync(path.join(options.cwd, 'test.js.snap'));
-		fs.unlinkSync(reportPath);
-	});
-
-	await exec.fixture(['--update-snapshots'], options);
-
-	const reportPath = path.join(options.cwd, 'test.js.md');
-
-	const report = fs.readFileSync(reportPath, {encoding: 'utf8'});
-	const ids = getSnapshotIds(report);
-
-	t.deepEqual(ids, [...ids].sort((a, b) => a - b));
-});
-
-test('deterministic and sorted over a large, random test case', async t => {
-	const options = {
-		cwd: exec.cwd('randomness'),
-		env: {
-			AVA_FORCE_CI: 'not-ci'
-		}
-	};
-
-	const snapshotPath = path.join(options.cwd, 'test.js.snap');
-	const reportPath = path.join(options.cwd, 'test.js.md');
-
-	// Run test
-	await exec.fixture(['--update-snapshots'], options);
-
-	// Assert snapshot is unchanged
-	const snapshot = fs.readFileSync(snapshotPath);
-
-	t.snapshot(snapshot);
-
-	// Assert report is sorted
-	const report = fs.readFileSync(reportPath);
-	const ids = getSnapshotIds(report);
-
-	t.deepEqual(ids, [...ids].sort((a, b) => a - b));
-});
+const getSnapshotIds = require('./helpers/get-snapshot-ids');
 
 const unsortedSnapshotPath = path.join(__dirname, 'fixtures', 'backwards-compatibility', 'unsorted', 'test.js.snap');
 const unsortedReportPath = path.join(__dirname, 'fixtures', 'backwards-compatibility', 'unsorted', 'test.js.md');
