+++ test/snapshot-order/helpers/get-snapshot-ids.js
@@ -1,0 +1,14 @@
+function getSnapshotIds(report) {
+	function * matchAll(string, regexp) {
+		let match;
+		while ((match = regexp.exec(string)) !== null) {
+			yield match;
+		}
+	}
+
+	const ids = [...matchAll(report, /'index: ([-.\d]+)'/g)].map(match => Number(match[1]));
+
+	return ids;
+}
+
+module.exports = getSnapshotIds;
