+++ test/snapshot-order/snapshots/randomness.js.md
@@ -1,6 +1,6 @@
-# Snapshot report for `test/snapshot-order/test.js`
+# Snapshot report for `test/snapshot-order/randomness.js`
 
-The actual snapshot is saved in `test.js.snap`.
+The actual snapshot is saved in `randomness.js.snap`.
 
 Generated by [AVA](https://avajs.dev).
